enable sort
